packages/client/libs/design-decisions/src/schemas/loadSchemasFromPaths.ts
import { loadSchemaFromFile } from './private/loadSchemaFromFile';
import { loadSchemasFromDirectory } from './private/loadSchemasFromDirectory';
import { SchemaData, SchemaId, SchemaMap } from './types';

export const loadSchemasFromPaths = async (paths: string[]): Promise<SchemaMap> => {
    const schemas = new Map<SchemaId, SchemaData>();
    for (const directoryPath of paths) {
        await loadSchemasFromDirectory(directoryPath, filePath =>
            loadSchemaFromFile(schemas, filePath),
        );
    }
    return schemas;
};
